extract layout header
