+++ components/local/FAQ/index.tsx
